+++ src/panoptes/utils/serial/device.py
@@ -6,9 +6,10 @@ from typing import Optional, Union, Callable
 
 import serial
 from loguru import logger
-from panoptes.utils import error
 from serial.threaded import LineReader, ReaderThread
 from serial.tools.list_ports import comports as get_comports
+
+from panoptes.utils import error
 
 
 @dataclass
@@ -231,7 +232,7 @@ class SerialDevice(object):
                 super(LineReader, this).connection_made(transport)
 
             def connection_lost(this, exc):
-                logger.warning(f'Disconnected from {self}')
+                logger.trace(f'Disconnected from {self}')
 
             def handle_line(this, data):
                 try: